Add check for package_dir being a URL
The package directory for an SVN build can be specified as a URL, but is_git()
does not like that and throws an exception. For now, just adding on a band-aid
patch that causes is_git to return False if given a URL (indicating that it's
not a git repo). In the future, we should allow building from a git URL.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-build/trunk@17370 4e558342-562e-0410-864c-e07659590f8c