Make koji and mock optional (SOFTWARE-2671)
Changes allow osg-build to continue working for tasks like rpmbuild,
even if `osgbuild.mock` or `osgbuild.kojiinter` are missing.

If a module is missing, options pertaining to that module will not show
up in `osg-build --help`, and it will direct the user to install the
package that contains the module. Attempting to invoke a task provided
by a missing module will cause an error and again direct the user to
install the package containing that module.