handle heavy-weight tags for upstream (SOFTWARE-3453)
if the expected hash matches the upstream ref, no need to dereference
them.  otherwise, dereference both (that is, until a non-tag object is
reached) and compare those.

(i am thinking that perhaps only the upstream ref needs to be
dereferenced, since it would be weird for our expected hash to refer to
a heavy-weight tag when the upstream ref does not; nevertheless i don't
really see a reason not to support this fallback.)