add fetch handler for 'uri' source type
this downloads a generic proto://... uri, and also implicitly adds a
file:// prefix for absolute paths.  (I am not wild about supporting it,
but it's supported in the old syntax, so for now we'll support it here
too.  Maybe later we'll decide to get rid of it.)

use chunked_read() to avoid reading the entire source file (of
arbitrary size) into memory at once

also compute a sha1sum of the file as we read it, rather than reading
it again after the fact

Note the line:

     if sha1sum: # or not ops.nocheck:

That "or not ops.nocheck" part should be uncommented once we want to
make sha1sums mandatory for uri/cached files also.  (Which i imagine
will be immediately after the upgraded osg-build is released, and
checksums are added to all our upstream.source files.)

Oh and there's more!

- allow specifying filename for uri source

a download uri might have either:

1. a basename that differs from the intended download filename
2. a query string that makes a mess of everything

and you might say 2. is just a specific instance of 1.

usually (but not always) the server sends a Content-Disposition header
with the intended attachment name in these cases, so that would be
one way to handle it, but it seems somehow cleaner to provide an
explicit 'filename' parameter to communicate exactly what was intended


(SOFTWARE-3568)