be a bit friendlier about deref_git_sha errors
git-rev-parse complains in an annoying and not really helpful way when
it doesn't know about the commit you're trying to dereference.

rather than put up with that, first verify that the object exists, and
then try to dereference.

if the object does not exist, log an error and just return the
un-dereferenced sha, rather than raising an Error.  (An Error will still
be raised later as a result, but this function is not the best place for
it.)


(SOFTWARE-3568)