un-document prefix and spec options for git types
Brian B has convinced me that there is no real use case for these -

If we control a git repo, we can make it conform; if we do not control a
git repo, we should probably be using type=cached rather than one of the
git types, since we don't control whether the upstream source stays or
goes away.

That leaves, perhaps, tool use for generating custom tarballs for
ourselves for arbitrary upstream git sources - but there does not seem
to be any good reason for using these options in our official builds.


(SOFTWARE-3568)