make docstring less grandiose
the 'like a python function call' description was supposed to describe
the unified nature of the old-style cached/uri type syntax with the
new-style meta-url syntax; but perhaps this serves more to confuse than
to enlighten, and users would be better served with the more practical
implications of the new syntax.

This is especially true now that implicit git types have been taken
off the table.  :shaking_fist_emoji:


(SOFTWARE-3568)