git-fetch compat for sl6 git 1.7.1 (SOFTWARE-3568)
turns out "git fetch tagname" doesn't work on git 1.7.1, which
is what runs on our SL6 Fermi VMs.  The EL6 git version on ingwe
is slightly newer and didn't suffer from this, but alas, not all
EL6's run the same git apparently.

Anyway, first try explicitly to fetch a tag object, but do it quietly
because it's ok if it fails - then fall back to a generic fetch, so
a branch name or HEAD can be specified.  If that one fails, error out.