fix git.get_branch (SOFTWARE-5476)
if 'git branch' output is:
```
* 3.6-upcoming
  devops
  osg-3.6
```

then `out.split()[-1]` will return `osg-3.6`, which is clearly wrong.

Instead, extract and the selected branch, marked by a '*'.

Raise an error if no branch is selected, or if the branch name has
a '(', which actually means that HEAD is detached, and not actually
on a branch.