Add osg-sign script
This is called with one of the signing keys from the signing_keys.ini
file, and one or more build NVRs; it will download the RPMs of those
builds, sign them with that key, and then import the newly signed RPMs
back into Koji.

Keys can be specified by name ("OSG-23-developer") or id ("92897c00").
You can also call the script with `--list-keys` to get a table of the known
keys as well as whether you have the corresponding secret key.

You can pass `--dry-run` to download and sign the RPMs only (and not import
them back) -- this doesn't require Koji permissions.
Importing an RPM signature requires the "sign" permission (or "admin").
The script checks if you have that permission first.

You can also pass `--results <DIR>` to keep the signed RPMs in the `<DIR>`
directory instead of deleting them after signing.

Having to pass specific NVRs is not very convenient; however, a separate
PR will include modifications to osg-promote to call out to this script
when promoting.