Use resources in osg-koji and osg-sign
On second thought, importlib_resources is available as an RPM
on EL 8, and we can use stdlib importlib.resources on EL 9.
Use it and avoid annoying deprecation warnings.

I don't yet use it to find promoter.ini because I want to avoid
merge conflicts with PRs I plan to make.