Add required key information to promoter.ini and Route object
required_keys is a new optional attribute for each route; it references the
keys in signing_keys.ini.  At least one of the keys listed must be used to sign
each package promoted using that route.

The key info in signing_keys.ini lists which distro versions a key can be used
to sign for (for example "OSG-4" can only be used on el9).  The actual key
required for a given build on a route is the intersection of the required_keys
for that route and the dver of that build.  I added a required_keys_for_dver()
method to Route for getting that intersection; this meant I had to turn Route
from a namedtuple to an object.  As a bonus, I get better type checking.