Add rejection reason for missing signature - sign packages in osg-promote (1/5)
This adds a new rejection reason to the Reject object of promoter.py.
This revamps the Reject object to add a `details` field, which is a dict
of extra information about the rejection.  This is because a 'missing
signature' rejection has details about the relevant RPMs and signing
keys, whereas a 'no matching build for dist' rejection has details about
the distro.

This changes the signature of the Reject constructor object from
`(pkg_or_build, dist, reason)` to `(pkg_or_build, reason, details)`