Verify RPM signatures before promoting - sign packages in osg-promote (3/5)
This adds a validate_signatures() check to adding a promotion.
For this, the Promoter needs the signing keys (as a dict of name to
SigningKey objects).  Like with the other rejections, you can explicitly
ignore this one by passing `ignore_signatures=True` to Promoter.add_promotion().
This is exposed as the command-line flag `--ignore-signatures`.

Each promotion route may have multiple keys in its required_keys list --
each RPM getting promoted must be signed with one of the keys in that list.
Since not all keys are usable for all distro versions (our 3.6 EL7/EL8
key is not usable for EL9 for example), the key that a build must be
signed by is the intersection of keys allowed for that route with keys that
work on that dver.

Actually validating the signatures is annoying: a GPG key may have multiple
subkeys and it's possible that the key id that we have listed in
signing_keys.ini is not the subkey that KojiHelper.get_rpms_and_keyids_in_build()
reports to us as having been used, so check against the other subkeys too.