Try to sign RPMs on promotion - sign packages in osg-promote (4/5)
Add an option to try and sign the RPMs of a build we're trying to promote
if it doesn't have the right signature and we have the right key.  By
default we try to sign if we're running interactively, but `--sign` and
`--no-sign` control that.

Each signing key is only tried once; we're assuming if it failed to sign
once, it's not going to work a second time, and try a different key (if
we can).  TODO: this doesn't distinguish between a failure to sign due to,
for example, not having the correct key, versus not being able to run
`osg-koji` or not having the necessary Koji permissions to import new
signatures.